requirements.txt
faker==19.12.0
python-dotenv==1.0.0